+++ source/Handlers/AutoCreateRedirectOnMove.cs
@@ -1,4 +1,5 @@
-﻿using Sitecore.Data;
+﻿using SharedSource.RedirectModule.Helpers;
+using Sitecore.Data;
 using Sitecore.Data.Events;
 using Sitecore.Data.Items;
 using Sitecore.Diagnostics;
@@ -8,7 +9,7 @@ using Sitecore.SecurityModel;
 using System;
 using System.Linq;
 
-namespace SharedSource.RedirectModule
+namespace SharedSource.RedirectModule.Handlers
 {
     public class AutoCreateRedirectOnMove
     {
